Fix discount BigDecimal initialization
The PayuIn#amount_ok? method initializes a BigDecimal with the discount as given in the params.
If the discount value is invalid as an argument for BigDecimal (e.g. '0.'), then the latest BigDecimal v1.3.2 raises an exception.
The behaviour of the current version of BigDecimal is to silently initialize any invalid arguments to a value of 0.00.

To support the new BigDecimal and preserve current behaviour, the method now uses `discount.to_d` to create a BigDecimal instead of `BigDecimal.new(discount.to_s)`, for the desired zero coercion.